refactor(book): move setupBuiltinFunctions logic to applyBuiltinFunctions
